stock_analysis/requirements.txt
flask==2.0.1
twstock==1.3.1
tqdm==4.62.3
requests==2.26.0
beautifulsoup4==4.10.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0